requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
blis==0.7.9
catalogue==2.0.8
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
datadog==0.47.0
detoxify==0.5.1
emfdscore @ git+https://github.com/medianeuroscience/emfdscore.git@64dfaea0cdb00f11921dc6278d46201d73128a80
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
ffmpeg-python==0.2.0
filelock==3.10.0
Flask==2.2.3
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
frozenlist==1.3.3
ftputil==5.0.4
future==0.18.3
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.13.2
hyperframe==5.2.0
idna==2.10
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mysql-connector-python==8.0.33
networkx==3.0
nltk==3.8.1
numba==0.57.0
numpy==1.24.2
openai==0.27.7
openai-whisper @ git+https://github.com/openai/whisper.git@248b6cb124225dd263bb9bd32d060b6517e067f8
opencv-python==4.8.0.74
packaging==23.0
pandas==1.5.3
pathy==0.10.1
preshed==3.0.8
progressbar2==4.2.0
protobuf==3.20.3
psutil==5.9.4
pydantic==1.9.2
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-utils==3.6.0
pytube @ git+https://github.com/pytube/pytube@a32fff39058a6f7e5e59ecd06a7467b71197ce35
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.97
sftputil==1.1.1
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
spacy==3.4.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
sympy==1.11.1
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.12.1
torch==2.0.0
tqdm==4.65.0
transformers==4.22.1
typer==0.4.2
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==0.10.1
Werkzeug==2.2.3
yarl==1.9.1
youtube-transcript-api==0.6.0
